requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
boxsdk>=3.9.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=7.4.0
httpx>=0.25.0
fastmcp>=0.9.0